requirements.txt
Django==2.2.3
django-bootstrap4==0.0.8
djangorestframework==3.10.3
gunicorn==20.0.4
pytz==2019.1
sqlparse==0.3.0
